pyproject.toml
[project]
name = "testx5"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django==5.2.11",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.11",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "isort>=8.0.0",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "test_x5.settings_test"
python_files = "tests.py test_*.py *_tests.py"

markers = []

log_cli = 1

filterwarnings = [
    "ignore::pytest.PytestWarning"
]

addopts = "-v"

testpaths = [
    "tests"
]

[tool.black]
line-length = 128
target-version = ["py312"]
exclude = '''
/(
  | migrations
)/
'''

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]
no_lines_before = "LOCALFOLDER"
line_length = 128
force_single_line = false
from_first = false
order_by_type = false
